+++ game/init.h
@@ -19,8 +19,6 @@
 namespace Verse
 {
     void init(Scene &scene, Config &config) {
-        scene.setSize(Vec2(256, 180));
-        
         EntityID test = scene.createEntity("Player");
         
         Component::Collider* collider = scene.addComponent<Component::Collider>(test);
@@ -39,7 +37,7 @@ namespace Verse
         animation->size = 26; //calculate_size
         
         Component::Actor* actor = scene.addComponent<Component::Actor>(test);
-        actor->controller = [&scene, test]() -> bool {return Controller::Player::controller(scene, Entity::getIndex(test));};
+        actor->controller = [&scene, test]() -> bool {return Controller::Player::controller(scene, test);};
         actor->max_move_speed = 100;
         actor->max_fall_speed = 250;
         actor->acc_ground = 1000;
@@ -52,47 +50,5 @@ namespace Verse
         Component::Camera* camera = scene.addComponent<Component::Camera>(test);
         System::Camera::init(camera, Vec2f(20, 90), config.camera_focus_size.to_float());
         System::Camera::setActive(camera, scene);
-        
-        
-        EntityID test1 = scene.createEntity("Luz");
-        Component::Light* light2 = scene.addComponent<Component::Light>(test1);
-        light2->pos = Vec2(120, 120);
-        light2->radius = 48;
-        
-        EntityID test2 = scene.createEntity("Luz");
-        Component::Light* light3 = scene.addComponent<Component::Light>(test2);
-        light3->pos = Vec2(200, 160);
-        light3->radius = 48;
-        
-        
-        EntityID test3 = scene.createEntity("Ground");
-        
-        Component::Collider* collider2 = scene.addComponent<Component::Collider>(test3);
-        collider2->transform = Rect2(192, 72, 8, 8);
-        
-        Component::Texture* texture2 = scene.addComponent<Component::Texture>(test3);
-        Graphics::Texture::loadTexture("res/graphics/ground2.png", texture2);
-        texture2->transform = Rect2(192, 72, 8, 8);
-        
-        
-        EntityID bg = scene.createEntity("Background");
-         
-        Component::Tilemap* bg_tilemap = scene.addComponent<Component::Tilemap>(bg);
-        bg_tilemap->tiles = System::Tilemap::load("res/levels/bg.png");
-        Graphics::Texture::loadTexture("res/graphics/background.png", bg_tilemap);
-        bg_tilemap->tex_size = Vec2(8, 8);
-        bg_tilemap->pos = Vec2(-56, -32);
-        
-        
-        EntityID tile = scene.createEntity("Tilemap");
-        
-        Component::Tilemap* tilemap = scene.addComponent<Component::Tilemap>(tile);
-        tilemap->tiles = System::Tilemap::load("res/levels/1.png");
-        Graphics::Texture::loadTexture("res/graphics/ground2.png", tilemap);
-        tilemap->tex_size = Vec2(8, 8);
-        tilemap->pos = Vec2(0, 96);
-        
-        Component::Collider* tile_collider = scene.addComponent<Component::Collider>(tile);
-        tile_collider->transform = Rect2(tilemap->pos, System::Tilemap::calculateSize(tilemap));
     }
 }